+++ apps/dot-components-e2e/src/integration/Dialog/Dialog.spec.ts
@@ -20,7 +20,7 @@ describe('dot-components: Dialog component', () => {
 
     it('dialog sizing', () => {
       cy.get('div.MuiDialog-paper')
-        .should('have.css', 'max-width', '600px')
+        .should('have.css', 'max-width', '720px')
         .and('have.css', 'margin-left', '32px');
     });